refactor(feedback): Rename `SentryFeedbackWidget` to `SentryFeedbackForm`
Renames SentryFeedbackWidget to SentryFeedbackForm while keeping a deprecated alias for backward compatibility.
