+++ packages/flutter/lib/sentry_flutter.dart
@@ -6,7 +6,8 @@ export 'package:sentry/sentry.dart';
 
 export 'src/binding_wrapper.dart'
     show BindingWrapper, SentryWidgetsFlutterBinding;
-export 'src/feedback/sentry_feedback_widget.dart';
+export 'src/feedback/sentry_feedback_form.dart';
+export 'src/feedback/sentry_feedback_widget.dart' show SentryFeedbackWidget;
 export 'src/flutter_sentry_attachment.dart';
 export 'src/integrations/load_release_integration.dart';
 export 'src/integrations/on_error_integration.dart';
